dto dau buoi re rach

+++ src/main/java/com/example/doctorcare/api/repository/CustomerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerRepository extends CrudRepository<CustomersEntity,Integer> {
+public interface CustomerRepository extends CrudRepository<CustomersEntity,Long> {
 }
