+++ src/main/java/com/example/doctorcare/api/service/UserRoleService.java
@@ -13,7 +13,7 @@ public class UserRoleService {
     @Autowired
     UserRoleRepository userRoleRepository;
 
-    public Optional<UserRoleEntity> findById(int id){
+    public Optional<UserRoleEntity> findById(Long id){
         return userRoleRepository.findById(id);
     }
 }
